requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via
    #   channels
    #   channels-redis
    #   django
    #   django-cors-headers
channels==4.2.2
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.2.1
    # via -r requirements.in
click==8.1.8
    # via face-recognition
cmake==3.31.4
    # via -r requirements.in
colorama==0.4.6
    # via click
dj-database-url==2.3.0
    # via -r requirements.in
django==5.1.6
    # via
    #   -r requirements.in
    #   channels
    #   dj-database-url
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==4.7.0
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.0
    # via -r requirements.in
dlib==19.24.6
    # via
    #   -r requirements.in
    #   face-recognition
face-recognition==1.3.0
    # via -r requirements.in
face-recognition-models==0.3.0
    # via face-recognition
gunicorn==23.0.0
    # via -r requirements.in
msgpack==1.1.0
    # via channels-redis
numpy==2.2.2
    # via
    #   -r requirements.in
    #   face-recognition
    #   opencv-python
opencv-python==4.11.0.86
    # via -r requirements.in
packaging==24.2
    # via gunicorn
pillow==11.1.0
    # via face-recognition
psycopg[binary]==3.2.4
    # via -r requirements.in
psycopg-binary==3.2.4
    # via psycopg
pyjwt==2.9.0
    # via djangorestframework-simplejwt
python-dotenv==1.0.1
    # via -r requirements.in
redis==5.2.1
    # via channels-redis
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   psycopg
tzdata==2025.1
    # via
    #   django
    #   psycopg
